requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.1
psycopg2==2.9.8
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
whitenoise==6.6.0